pyproject.toml
[tool.poetry]
name = "owa-discordbot"
version = "0.1.0"
description = "Custom bot for various fun activities!"
authors = ["rozium"]

[tool.poetry.dependencies]
python = "^3.6"
Flask = "^1.1.2"
PyNaCl = "^1.4.0"
dnspython = "^2.0.0"
"discord.py" = "^1.5.1"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
